pyproject.toml
[project]
name = "todoapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.3.0",
    "fastapi>=0.128.0",
    "jinja2>=3.1.6",
    "motor>=3.7.1",
    "passlib>=1.7.4",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "slowapi>=0.1.9",
    "uvicorn[standard]>=0.40.0",
]
